+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Upcoming
+- Add `.destroy` option
 
 ## 1.0.0
 - mechanism working